Move utility functions to utils
This commit moves randomString and formatComplex to utils.
