+++ packages/qiskit-sim/lib/utils.js
@@ -10,9 +10,34 @@
 'use strict';
 
 const utils = require('@qiskit/utils');
+const math = require('mathjs');
 
 const pkgName = require('../package.json').name;
 
 utils.dbg = fullPath => utils.debug(`${pkgName}:${utils.pathToTag(fullPath)}`);
+
+utils.randomString = length => {
+  const len = length || 17;
+  let text = '';
+  // var first char to be letter
+  let charset = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
+
+  text += charset.charAt(Math.floor(Math.random() * charset.length));
+  // other chars can be numbers
+  charset += '0123456789';
+
+  for (let i = 0; i < len; i += 1) {
+    text += charset.charAt(Math.floor(Math.random() * charset.length));
+  }
+
+  return text;
+};
+
+utils.formatComplex = complex => {
+  const re = math.round(complex.re, 8);
+  const im = math.round(complex.im, 8);
+
+  return `${re}${im >= 0 ? '+' : '-'}${math.abs(im)}i`;
+};
 
 module.exports = utils;
